Allow unsupported ABIs in CI to work around size-of dependency issue
Added RUSTFLAGS to allow unsupported_calling_conventions as warnings instead of errors.

This resolves the catch-22:
- Rust 1.83: Can't compile base64ct (needs edition 2024)
- Rust 1.84+: Can't compile size-of (ABI hard errors)

Solution: Use stable Rust with -A unsupported_calling_conventions flag.
This is safe since size-of is only a transitive dependency and the ABI
issues don't affect our actual code.

See KNOWN_ISSUES.md for details on the upstream size-of issue.