Override RUSTFLAGS in CI to allow size-of ABI errors
Set RUSTFLAGS="" in job env to override the default -D warnings that
causes size-of v0.1.5 ABI errors to fail the build.

The size-of crate uses unsupported ABIs (aapcs, stdcall, fastcall) that
are now hard errors in Rust 1.84+. By clearing RUSTFLAGS, we allow these
to compile while still enforcing Clippy warnings via -- -D warnings.

This is safe as size-of is only a transitive dependency from starknet-types-core.